#14 #30 add components for AssetEventTable
I let this be a component to handle more of the functions within the component - I'm thinking we could potentially add some "extra function" if needed. Like fex - afterClose or such

put some types in a file assettypes so we can specificly extend from them in components - they already look very alike.

added a type for assetid which as I see can be both a string and a number. Referred it in where I saw it was in use

created storybook for assedDetailsPanel as it's now called.