Component asset event table (#41)
* #14 #30 add components for AssetEventTable

I let this be a component to handle more of the functions within the component - I'm thinking we could potentially add some "extra function" if needed. Like fex - afterClose or such

put some types in a file assettypes so we can specificly extend from them in components - they already look very alike.

added a type for assetid which as I see can be both a string and a number. Referred it in where I saw it was in use

created storybook for assedDetailsPanel as it's now called.

* remove unused refs

* #14 #30 add components for AssetEventTable

I let this be a component to handle more of the functions within the component - I'm thinking we could potentially add some "extra function" if needed. Like fex - afterClose or such

put some types in a file assettypes so we can specificly extend from them in components - they already look very alike.

added a type for assetid which as I see can be both a string and a number. Referred it in where I saw it was in use

created storybook for assedDetailsPanel as it's now called.

* #14 #28 add snapshots for test

* #14 add some references to components

* #14 give describe test correct name

* Removed obsolute snapshot

* #14 correct naming assetdetaoslpaneltest

* Make sure timestamps is in UTC timezone