feat: Add optional AutoPaging prop to AssetTree
This allows us to override the default 25 asset limit in the SDK when
fetching sub-assets in the view, by passing an optional
`autoPagingToArrayOption` prop with an `{ limit: 123 }` object.

This added flexibility might come at the cost of performance if one
passes large limits, or if one uses the supported `-1` that lists
however many entries are in the sub-tree.