fix(tenant-selector): remove confusing "x" on invalid tenant
When the tenant is invalid, the UI displays an invalid state with
- an error message
- the input box highlighted in red
- a red X inside the input box

This is antd's default behaviour but it is confusing for users that there is an X inside the input field, which does nothing. We may expect that it should be a clickable button to clear out the invalid input.

Would like to just remove it to avoid confusion.

Fixes [OPSUP-388]