feat(svg-viewer): use cogs.js buttons on the svg viewer
Both OpSup and ADI are switching our applications buttons to use the cogs.js standard rather than antd directly.
The SVGViewer is the only place where the buttons are still, now, inconsistent with the rest of the application.. so it would be nice to update them also.

I believe the original intention was for gearbox to use cogs.js styles where they are available anyway?

Reference:
https://github.com/cognitedata/operational-intelligence/pull/4244
https://github.com/cognitedata/operations-support/pull/3092

Fixes [OPSUP-1039]